Support for Chrome Custom Tabs on Android
